DI 변경
Factory -> viewModel
RepositoryImpl DI 적용
