+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:theme="@style/Theme.UserList"
         tools:targetApi="31">
         <activity
-            android:name=".ui.view.main.MainActivity"
+            android:name=".view.main.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.UserList.NoActionBar">
