+++ app/src/main/java/org/project/userlist/App.kt
@@ -3,30 +3,34 @@ package org.project.userlist
 import android.app.Application
 import org.koin.android.ext.koin.androidContext
 import org.koin.android.ext.koin.androidLogger
-import org.koin.core.context.GlobalContext.startKoin
+import org.koin.androidx.viewmodel.dsl.viewModel
+import org.koin.core.context.startKoin
 import org.koin.dsl.module
 import org.project.userlist.data.local.UsersDb
-import org.project.userlist.data.repository.UsersRepository
-import org.project.userlist.ui.view.userList.UsersViewModel
+import org.project.userlist.view.userList.UsersViewModel
 import org.project.userlist.data.network.Retrofit
-import org.project.userlist.ui.view.userList.BookMarkUsersViewModel
+import org.project.userlist.data.repository.UsersRepository
+import org.project.userlist.data.repository.UsersRepositoryImpl
+import org.project.userlist.view.userList.BookMarkUsersViewModel
 
 class App :Application() {
 
     override fun onCreate() {
         super.onCreate()
+
         startKoin {
             androidLogger()
             androidContext(this@App)
             modules(appModule)
         }
     }
+
     private val appModule = module {
         single { UsersDb.create(context = androidContext()) }
         single { Retrofit.instance }
-        single { UsersRepository(retrofitApi = get(), db = get()) }
+        single<UsersRepository> { UsersRepositoryImpl(retrofitApi = get(), db = get()) }
 
-        factory { UsersViewModel(usersRepository = get()) }
-        factory { BookMarkUsersViewModel(usersRepository = get()) }
+        viewModel { UsersViewModel(usersRepository = get()) }
+        viewModel { BookMarkUsersViewModel(usersRepository = get()) }
     }
 }
