+++ app/src/main/java/org/project/userlist/view/userList/UsersViewModel.kt
@@ -1,7 +1,6 @@
-package org.project.userlist.ui.view.userList
+package org.project.userlist.view.userList
 
 import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import androidx.paging.PagedList
@@ -33,12 +32,12 @@ class UsersViewModel(
 
     // 북마크 체크 시 BookMarkUsers에 해당 Users 데이터 저장
     suspend fun insertBookMarkUsers(users:Users) {
-        usersRepository.insertBookMarkUsersFromUsers(users)
+        usersRepository.insertBookMarkUsers(users)
     }
 
     // 북마크 체크 해제 시 BookMarkUsers에 해당 Users 데이터 삭제
     suspend fun deleteBookMarkUsers(users:Users) {
-        usersRepository.deleteBookMarkUsersFromUsers(users)
+        usersRepository.deleteBookMarkUsers(users)
     }
 
     // 온라인(mobile, wifi 연결)으로 전환 시 재시도
@@ -51,7 +50,7 @@ class UsersViewModel(
     // 다음 페이지 로드 재시도
     fun reTryListener(connected:Boolean) {
         viewModelScope.launch(Dispatchers.IO) {
-            usersRepository.reTryListener(connected)
+            usersRepository.reTry(connected)
         }
     }
 }
