+++ app/src/main/res/layout/fragment_bookmark_users.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.view.userList.BookmarkUserFragment">
+    tools:context=".view.userList.BookmarkUserFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/book_mark_recycler_view"
