+++ app/src/main/res/layout/fragment_users.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.view.userList.UsersFragment">
+    tools:context=".view.userList.UsersFragment">
 
 
     <androidx.recyclerview.widget.RecyclerView
